requirements.txt
fastapi
uvicorn
jinja2
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-mgmt-resource
azure-mgmt-subscription
psycopg2-binary
python-dotenv
